requirements.txt
streamlit==1.17.0
nltk
advertools
numpy
pystempel
sklearn
scikit-learn
streamlit
spacy
transformers
sentence-transformers
sentencepiece
torch
torchvision
xlsxwriter